Improved legibility

+++ app/Types/Armor.php
@@ -10,27 +10,27 @@ class Armor extends Translate implements Item
     public function getStructure(): array
     {
         return [
-            'level' => 'int',
-            'class' => 'int',
-            'strength' => 'int',
-            'vitality' => 'int',
-            'dexterity' => 'int',
-            'magic' => 'int',
+            'level' => 'int32',
+            'class' => 'int32',
+            'strength' => 'int32',
+            'vitality' => 'int32',
+            'dexterity' => 'int32',
+            'magic' => 'int32',
             'min_durability' => 'durability',
             'max_durability' => 'durability',
-            'item_type' => 'int',
+            'item_type' => 'int32',
             'item_flag' => 'short',
             'name_length' => 'name_length',
             'name' => 'name',
-            'physical_defense' => 'int8',
-            'evasion' => 'int8',
-            'mp' => 'int8',
-            'hp' => 'int8',
-            'metal_resistance' => 'int8',
-            'wood_resistance' => 'int8',
-            'water_resistance' => 'int8',
-            'fire_resistance' => 'int8',
-            'earth_resistance' => 'int8',
+            'physical_defense' => 'int64',
+            'evasion' => 'int64',
+            'mp' => 'int64',
+            'hp' => 'int64',
+            'metal_resistance' => 'int64',
+            'wood_resistance' => 'int64',
+            'water_resistance' => 'int64',
+            'fire_resistance' => 'int64',
+            'earth_resistance' => 'int64',
             'sockets_count' => 'sockets_count',
             'sockets' => 'sockets',
             'addons_count' => 'addons_count',
@@ -38,7 +38,7 @@ class Armor extends Translate implements Item
         ];
     }
 
-    public function getMinimumLength() : int
+    public function getMinimumLength() : int32
     {
         return 135;
     }
